+++ src/global/before.js
@@ -1,3 +1,4 @@
+const filterStack = require('../filter_stack');
 const Example = require('../example');
 
 module.exports = {
@@ -16,7 +17,10 @@ module.exports = {
       for(let i=0; i< this.beforeHooks.length; i++) {
         const hook = this.beforeHooks[i];
         if (!hook.hasRun) {
-          await hook.run().catch(error => hook.error = error);
+          await hook.run().catch(error => {
+            filterStack(error);
+            hook.error = error;
+          });
           hook.hasRun = true;
         }
         if (hook.error) throw hook.error;