+++ src/index.js
@@ -1,6 +1,6 @@
-// set up stack filtering
-require('./error');
+'use strict';
 
+require('./filter_stack'); // needs to have 'original' stack prep stored
 const RunnerManager = require('./runner_manager');
 const Options = require('./options');